Fix source references
We use the source references from the parse data. This has the
advantage that we don't need to evaluate the package code
twice.

Right now we only match top level functions from the package
namespace, that we created with the usual `fun <- function()...`
expression. With this approach it is hard to match other objects.

Closes #3.