Instrument top-level expressions in R code
I.e. the ones that run at install time. This marks at
`function` line and the final `}` of every function as covered.
It makes sense, but I am not sure if I like it. Maybe we can
mark it as "covered at install time". In general we could
differentiate between running at install time and running at
run time.

Closes #4, although I might undo or refine this.